requirements.txt
matplotlib
openai
pandas
python-dotenv
requests
seaborn
tenacity
pytest
